fix rubocop linters issue
